requirements.txt
flask>=2.3.0
requests>=2.31.0
pyyaml>=6.0
user-agents>=2.2.0
inquirerpy
rich>=13.7.0
pyserial
